Redesign: move Api.js outside of components folder.
