+++ frontend/js/components/App.react.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import Api from './Api';
+import Api from '../Api';
 import Contact from './Contact.react';
 import ContactForm from './ContactForm.react';
 import React, { Component } from "react";
